[RENAME]  : ✏️ GlobalStyle 위치 변경에 따른 import 경로 변경
