+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Router from './shared/Router';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import GlobalStyle from './shared/GlobalStyle';
+import GlobalStyle from './styles/GlobalStyle';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 
 const queryClient = new QueryClient({
