+++ src/components/Card/style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLOR } from '../../constants/colors';
+import { COLOR } from '../../styles/colors';
 
 export const StCardBlock = styled.div`
   background-color: ${COLOR.PAGE_LIGHTBLUE};
