+++ src/components/UploadedFileTab/UploadedOne/style.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import { COLOR } from '../../../constants/colors';
+import { COLOR } from '../../../styles/colors';
 
 export const StUploadedFileBlock = styled.div`
   width: 100%;
